feat(output): adds HTML formatting service
Adds html_formatter as a public service, moves related
API inside the service.

Adds public API to prepare an HTML string without having to
rely on output/longtext view, which is not flexible and adds
additional markup.
Adds a hook to allow plugins to implement custom preparation logic
