+++ engine/classes/Elgg/Di/ServiceProvider.php
@@ -29,99 +29,100 @@ use Zend\Mail\Transport\TransportInterface as Mailer;
  * IDEs to auto-complete properties and understand the types returned. Extension allows us to keep
  * the container generic.
  *
- * @property-read \Elgg\Database\AccessCollections         $accessCollections
- * @property-read \ElggCache                               $accessCache
- * @property-read \Elgg\ActionsService                     $actions
- * @property-read \Elgg\Users\Accounts                     $accounts
- * @property-read \Elgg\Database\AdminNotices              $adminNotices
- * @property-read \Elgg\Ajax\Service                       $ajax
- * @property-read \Elgg\Amd\Config                         $amdConfig
- * @property-read \Elgg\Database\AnnotationsTable          $annotationsTable
- * @property-read \ElggAutoP                               $autoP
- * @property-read \Elgg\AutoloadManager                    $autoloadManager
- * @property-read \Elgg\BootService                        $boot
- * @property-read \Elgg\Application\CacheHandler           $cacheHandler
- * @property-read \Elgg\Assets\CssCompiler                 $cssCompiler
- * @property-read \Elgg\Security\Csrf                      $csrf
- * @property-read \Elgg\ClassLoader                        $classLoader
- * @property-read \Elgg\Cli                                $cli
- * @property-read \Symfony\Component\Console\Input\ArgvInput $cli_input
+ * @property-read \Elgg\Database\AccessCollections                $accessCollections
+ * @property-read \ElggCache                                      $accessCache
+ * @property-read \Elgg\ActionsService                            $actions
+ * @property-read \Elgg\Users\Accounts                            $accounts
+ * @property-read \Elgg\Database\AdminNotices                     $adminNotices
+ * @property-read \Elgg\Ajax\Service                              $ajax
+ * @property-read \Elgg\Amd\Config                                $amdConfig
+ * @property-read \Elgg\Database\AnnotationsTable                 $annotationsTable
+ * @property-read \ElggAutoP                                      $autoP
+ * @property-read \Elgg\AutoloadManager                           $autoloadManager
+ * @property-read \Elgg\BootService                               $boot
+ * @property-read \Elgg\Application\CacheHandler                  $cacheHandler
+ * @property-read \Elgg\Assets\CssCompiler                        $cssCompiler
+ * @property-read \Elgg\Security\Csrf                             $csrf
+ * @property-read \Elgg\ClassLoader                               $classLoader
+ * @property-read \Elgg\Cli                                       $cli
+ * @property-read \Symfony\Component\Console\Input\ArgvInput      $cli_input
  * @property-read \Symfony\Component\Console\Output\ConsoleOutput $cli_output
- * @property-read \Elgg\Cron                               $cron
- * @property-read \ElggCrypto                              $crypto
- * @property-read \Elgg\Config                             $config
- * @property-read \Elgg\Database\ConfigTable               $configTable
- * @property-read \Elgg\Cache\DataCache                    $dataCache
- * @property-read \Elgg\Database                           $db
- * @property-read \Elgg\Database\DbConfig                  $dbConfig
- * @property-read \Elgg\DeprecationService                 $deprecation
- * @property-read \Elgg\DI\PublicContainer                 $dic
- * @property-read \Di\ContainerBuilder                     $dic_builder
- * @property-read \Elgg\Di\DefinitionCache                 $dic_cache
- * @property-read \Elgg\Di\DefinitionLoader                $dic_loader
- * @property-read \Elgg\EmailService                       $emails
- * @property-read \Elgg\Cache\EntityCache                  $entityCache
- * @property-read \Elgg\EntityPreloader                    $entityPreloader
- * @property-read \Elgg\Database\EntityTable               $entityTable
- * @property-read \Elgg\EventsService                      $events
- * @property-read \Elgg\Assets\ExternalFiles               $externalFiles
- * @property-read \ElggCache                               $fileCache
- * @property-read \ElggDiskFilestore                       $filestore
- * @property-read \Elgg\FormsService                       $forms
- * @property-read \Elgg\Gatekeeper                         $gatekeeper
- * @property-read \Elgg\HandlersService                    $handlers
- * @property-read \Elgg\Security\HmacFactory               $hmac
- * @property-read \Elgg\PluginHooksService                 $hooks
- * @property-read \Elgg\EntityIconService                  $iconService
- * @property-read \Elgg\ImageService                       $imageService
- * @property-read \Elgg\Invoker                            $invoker
- * @property-read \Elgg\Logger                             $logger
- * @property-read Mailer                                   $mailer
- * @property-read \Elgg\Menu\Service                       $menus
- * @property-read \Elgg\Cache\MetadataCache                $metadataCache
- * @property-read \Elgg\Database\MetadataTable             $metadataTable
- * @property-read \Elgg\Database\Mutex                     $mutex
- * @property-read \Elgg\Notifications\NotificationsService $notifications
- * @property-read \Elgg\PasswordService                    $passwords
- * @property-read \Elgg\PersistentLoginService             $persistentLogin
- * @property-read \Elgg\Database\Plugins                   $plugins
- * @property-read \ElggCache                               $privateSettingsCache
- * @property-read \Elgg\Database\PrivateSettingsTable      $privateSettings
- * @property-read \Elgg\Application\Database               $publicDb
- * @property-read \Elgg\Database\QueryCounter              $queryCounter
- * @property-read \Elgg\RedirectService                    $redirects
- * @property-read \Elgg\Http\Request                       $request
- * @property-read \Elgg\Router\RequestContext              $requestContext
- * @property-read \Elgg\Http\ResponseFactory               $responseFactory
- * @property-read \Elgg\Database\RelationshipsTable        $relationshipsTable
- * @property-read \Elgg\Router\RouteCollection             $routeCollection
- * @property-read \Elgg\Router\RouteRegistrationService    $routes
- * @property-read \Elgg\Router                             $router
- * @property-read \Elgg\Database\Seeder                    $seeder
- * @property-read \Elgg\Application\ServeFileHandler       $serveFileHandler
- * @property-read \ElggSession                             $session
- * @property-read \Elgg\Cache\SessionCache                 $sessionCache
- * @property-read \Elgg\Search\SearchService               $search
- * @property-read \Elgg\Cache\SimpleCache                  $simpleCache
- * @property-read \Elgg\Database\SiteSecret                $siteSecret
- * @property-read \Elgg\Forms\StickyForms                  $stickyForms
- * @property-read \Elgg\Cache\SystemCache                  $systemCache
- * @property-read \Elgg\SystemMessagesService              $systemMessages
- * @property-read \Elgg\Views\TableColumn\ColumnFactory    $table_columns
- * @property-read \ElggTempDiskFilestore                   $temp_filestore
- * @property-read \Elgg\Timer                              $timer
- * @property-read \Elgg\I18n\Translator                    $translator
- * @property-read \Elgg\Security\UrlSigner                 $urlSigner
- * @property-read \Elgg\UpgradeService                     $upgrades
- * @property-read \Elgg\Upgrade\Locator                    $upgradeLocator
- * @property-read \Elgg\Router\UrlGenerator                $urlGenerator
- * @property-read \Elgg\Router\UrlMatcher                  $urlMatcher
- * @property-read \Elgg\UploadService                      $uploads
- * @property-read \Elgg\UserCapabilities                   $userCapabilities
- * @property-read \Elgg\Database\UsersTable                $usersTable
- * @property-read \Elgg\ViewsService                       $views
- * @property-read \Elgg\Cache\ViewCacher                   $viewCacher
- * @property-read \Elgg\WidgetsService                     $widgets
+ * @property-read \Elgg\Cron                                      $cron
+ * @property-read \ElggCrypto                                     $crypto
+ * @property-read \Elgg\Config                                    $config
+ * @property-read \Elgg\Database\ConfigTable                      $configTable
+ * @property-read \Elgg\Cache\DataCache                           $dataCache
+ * @property-read \Elgg\Database                                  $db
+ * @property-read \Elgg\Database\DbConfig                         $dbConfig
+ * @property-read \Elgg\DeprecationService                        $deprecation
+ * @property-read \Elgg\DI\PublicContainer                        $dic
+ * @property-read \Di\ContainerBuilder                            $dic_builder
+ * @property-read \Elgg\Di\DefinitionCache                        $dic_cache
+ * @property-read \Elgg\Di\DefinitionLoader                       $dic_loader
+ * @property-read \Elgg\EmailService                              $emails
+ * @property-read \Elgg\Cache\EntityCache                         $entityCache
+ * @property-read \Elgg\EntityPreloader                           $entityPreloader
+ * @property-read \Elgg\Database\EntityTable                      $entityTable
+ * @property-read \Elgg\EventsService                             $events
+ * @property-read \Elgg\Assets\ExternalFiles                      $externalFiles
+ * @property-read \ElggCache                                      $fileCache
+ * @property-read \ElggDiskFilestore                              $filestore
+ * @property-read \Elgg\FormsService                              $forms
+ * @property-read \Elgg\Gatekeeper                                $gatekeeper
+ * @property-read \Elgg\HandlersService                           $handlers
+ * @property-read \Elgg\Security\HmacFactory                      $hmac
+ * @property-read \Elgg\Views\HtmlFormatter                       $html_formatter
+ * @property-read \Elgg\PluginHooksService                        $hooks
+ * @property-read \Elgg\EntityIconService                         $iconService
+ * @property-read \Elgg\ImageService                              $imageService
+ * @property-read \Elgg\Invoker                                   $invoker
+ * @property-read \Elgg\Logger                                    $logger
+ * @property-read Mailer                                          $mailer
+ * @property-read \Elgg\Menu\Service                              $menus
+ * @property-read \Elgg\Cache\MetadataCache                       $metadataCache
+ * @property-read \Elgg\Database\MetadataTable                    $metadataTable
+ * @property-read \Elgg\Database\Mutex                            $mutex
+ * @property-read \Elgg\Notifications\NotificationsService        $notifications
+ * @property-read \Elgg\PasswordService                           $passwords
+ * @property-read \Elgg\PersistentLoginService                    $persistentLogin
+ * @property-read \Elgg\Database\Plugins                          $plugins
+ * @property-read \ElggCache                                      $privateSettingsCache
+ * @property-read \Elgg\Database\PrivateSettingsTable             $privateSettings
+ * @property-read \Elgg\Application\Database                      $publicDb
+ * @property-read \Elgg\Database\QueryCounter                     $queryCounter
+ * @property-read \Elgg\RedirectService                           $redirects
+ * @property-read \Elgg\Http\Request                              $request
+ * @property-read \Elgg\Router\RequestContext                     $requestContext
+ * @property-read \Elgg\Http\ResponseFactory                      $responseFactory
+ * @property-read \Elgg\Database\RelationshipsTable               $relationshipsTable
+ * @property-read \Elgg\Router\RouteCollection                    $routeCollection
+ * @property-read \Elgg\Router\RouteRegistrationService           $routes
+ * @property-read \Elgg\Router                                    $router
+ * @property-read \Elgg\Database\Seeder                           $seeder
+ * @property-read \Elgg\Application\ServeFileHandler              $serveFileHandler
+ * @property-read \ElggSession                                    $session
+ * @property-read \Elgg\Cache\SessionCache                        $sessionCache
+ * @property-read \Elgg\Search\SearchService                      $search
+ * @property-read \Elgg\Cache\SimpleCache                         $simpleCache
+ * @property-read \Elgg\Database\SiteSecret                       $siteSecret
+ * @property-read \Elgg\Forms\StickyForms                         $stickyForms
+ * @property-read \Elgg\Cache\SystemCache                         $systemCache
+ * @property-read \Elgg\SystemMessagesService                     $systemMessages
+ * @property-read \Elgg\Views\TableColumn\ColumnFactory           $table_columns
+ * @property-read \ElggTempDiskFilestore                          $temp_filestore
+ * @property-read \Elgg\Timer                                     $timer
+ * @property-read \Elgg\I18n\Translator                           $translator
+ * @property-read \Elgg\Security\UrlSigner                        $urlSigner
+ * @property-read \Elgg\UpgradeService                            $upgrades
+ * @property-read \Elgg\Upgrade\Locator                           $upgradeLocator
+ * @property-read \Elgg\Router\UrlGenerator                       $urlGenerator
+ * @property-read \Elgg\Router\UrlMatcher                         $urlMatcher
+ * @property-read \Elgg\UploadService                             $uploads
+ * @property-read \Elgg\UserCapabilities                          $userCapabilities
+ * @property-read \Elgg\Database\UsersTable                       $usersTable
+ * @property-read \Elgg\ViewsService                              $views
+ * @property-read \Elgg\Cache\ViewCacher                          $viewCacher
+ * @property-read \Elgg\WidgetsService                            $widgets
  *
  * @access private
  */
@@ -391,6 +392,15 @@ class ServiceProvider extends DiContainer {
 
 		$this->setFactory('hmac', function(ServiceProvider $c) {
 			return new \Elgg\Security\HmacFactory($c->siteSecret, $c->crypto);
+		});
+
+		$this->setFactory('html_formatter', function(ServiceProvider $c) {
+			return new \Elgg\Views\HtmlFormatter(
+				$c->logger,
+				$c->views,
+				$c->hooks,
+				$c->autoP
+			);
 		});
 
 		$this->setFactory('hooks', function(ServiceProvider $c) {
